backend/pyproject.toml
[tool.poetry]
name = "gomingo"
version = "0.1.0"
description = "gomingo project"
authors = ["Dai Fukunaga <dai.pochi.700@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.1.0"
matplotlib = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
